reorganize components into common-app folder
