+++ common-app/components/charts/chart-card.component.tsx
