+++ common-app/components/charts/chart-page.component.tsx
