+++ layouts/chart-page.layout.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { makeStyles, Theme, createStyles, Typography } from "@material-ui/core";
 import { ChartInfoVm } from "core";
 import { AppLayout, HeaderLayout, FooterLayout, MainChartPageLayout } from ".";
-import { HeaderComponent, FooterComponent } from "components";
+import { HeaderComponent, FooterComponent } from "common-app/components";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
