+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { MainChartListContainer } from "pods/main-chart-list";
 import { AppLayout, HeaderLayout, FooterLayout, MainLayout } from "layouts";
-import { FooterComponent, HeaderComponent } from "components";
+import { FooterComponent, HeaderComponent } from "common-app/components";
 import { getAllCharts } from "core/api";
 
 export default function Index() {
